# backend/swagger/openapi.yaml
openapi: "3.0.0"

info:
  title: Todos
  version: "1.0"

servers:
  - url: /api/v1

components:
  parameters:
    PathId:
      name: id
      in: path
      schema:
        type: string
        format: uuid
      required: true

    IsCompleted:
      name: is_completed
      in: query
      schema:
        type: boolean
      required: false

    Limit:
      name: limit
      in: query
      schema:
        type: integer
      example: 10
      required: false

    Offset:
      name: offset
      in: query
      schema:
        type: integer
      example: 0
      required: false

  schemas:
    Credentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
      required:
        - username
        - password

    Token:
      type: object
      properties:
        access:
          type: string
        refresh:
          type: string
      required:
        - access
        - refresh

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        joined_at:
          type: string
          format: date-time
      required:
        - id
        - username
        - joined_at

    ChangePassword:
      type: object
      properties:
        password:
          type: string
          format: password
        new_password:
          type: string
          format: password
      required:
        - password
        - new_password

    Todo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        is_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - name
        - is_completed
        - created_at
        - updated_at
        - completed_at

    TodoList:
      type: array
      items:
        $ref: "#/components/schemas/Todo"

    TodoListPaginated:
      type: array
      items:
        type: object
        properties:
          data:
            $ref: "#/components/schemas/TodoList"
          count:
            type: integer
        required:
          - data
          - count

    NewTodo:
      type: object
      properties:
        name:
          type: string
      required:
        - name

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    UnauthorizedError:
      description: Unauthorized

    ForbiddenError:
      description: Forbidden

    NotFoundError:
      description: Not Found

    InternalServerError:
      description: Internal Server Error

tags:
  - name: auth
  - name: todos

paths:
  /sign-in:
    post:
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"

  /sign-up:
    post:
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"

  /sign-refresh:
    post:
      tags:
        - auth
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"

  /profile:
    get:
      tags:
        - auth
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /change-password:
    post:
      tags:
        - auth
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePassword"
      responses:
        "204":
          description: NO CONTENT

  /todos:
    get:
      tags:
        - todos
      parameters:
        - $ref: "#/components/parameters/IsCompleted"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoListPaginated"

    post:
      tags:
        - todos
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTodo"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"

  /todos/{id}:
    get:
      tags:
        - todos
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"

    patch:
      tags:
        - todos
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTodo"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"

    delete:
      tags:
        - todos
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "204":
          description: No Content

  /todos/{id}/complete:
    post:
      tags:
        - todos
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"

  /todos/{id}/revert:
    post:
      tags:
        - todos
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"